// yukka-ui/src/app/services/models/ustawienia.ts
/* tslint:disable */
/* eslint-disable */
export interface Ustawienia {
  galeriaOcenaKomentarze?: boolean;
  galeriaPokaz?: boolean;
  id?: number;
  ogrodOcenaKomentarze?: boolean;
  ogrodPokaz?: boolean;
  powiadomieniaKomentarzeGaleria?: boolean;
  powiadomieniaKomentarzeOdpowiedz?: boolean;
  powiadomieniaKomentarzeOgrod?: boolean;
  powiadomieniaOgrodKwitnienie?: boolean;
  powiadomieniaOgrodOwocowanie?: boolean;
  powiadomieniaOgrodPodlewanie?: boolean;
  statystykiProfilu?: boolean;
}