// yukka-ui/src/app/services/models/ustawienia-request.ts
/* tslint:disable */
/* eslint-disable */
export interface UstawieniaRequest {
  galeriaOcenaKomentarze: boolean;
  galeriaPokaz: boolean;
  ogrodOcenaKomentarze: boolean;
  ogrodPokaz: boolean;
  powiadomieniaKomentarzeGaleria: boolean;
  powiadomieniaKomentarzeOdpowiedz: boolean;
  powiadomieniaKomentarzeOgrod: boolean;
  powiadomieniaOgrodKwitnienie: boolean;
  powiadomieniaOgrodOwocowanie: boolean;
  powiadomieniaOgrodPodlewanie: boolean;
  statystykiProfilu: boolean;
}